refactor: add add to cart handler
